Use Principal Component Analysis(PCA) to solve the mnist classification problem
Using PCA to generate the most common variations of the data and then project
the validation set onto the basis of the eigenvectors. The classification is
done by using the number with the smallest residual after projection. Not sure
if this idea has any merit compared to e.g nearest neighbour or using clusters.