Added a terrible resampling classifier.
The resampling iteratively builds a set meant to represent a
certain class. It does this by adding the sample furthest away
from the nearest point in the current set. On it's own this
classifier is extremly sensitive to outliers. It is however a
pretty easy way to see very different instance of the same class.